# back/swagger.yaml
openapi: 3.0.0
info:
  title: DevFood API
  description: API para aplicação DevFood
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Servidor de desenvolvimento local
...
paths:
...
  /endereco:
    post:
      tags:
        - Endereços
      summary: Cadastrar novo endereço
      description: Adiciona um novo endereço para um usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  required:
                    - id_usuario
                    - attributes
                  properties:
                    id_usuario:
                      type: string
                      format: uuid
                      example: "550e8400-e29b-41d4-a716-446655440000"
                    attributes:
                      type: object
                      required:
                        - logradouro
                        - bairro
                        - cidade
                        - estado
                        - pais
                        - numero
                        - tipo
                      properties:
                        logradouro:
                          type: string
                          example: "Rua Exemplo"
                        bairro:
                          type: string
                          example: "Bairro Exemplo"
                        cidade:
                          type: string
                          example: "Cidade Exemplo"
                        estado:
                          type: string
                          example: "UF"
                        pais:
                          type: string
                          example: "País Exemplo"
                        numero:
                          type: integer
                          example: 123
                        complemento:
                          type: string
                          example: "Apto 123"
                        tipo:
                          type: string
                          enum: ["casa", "trabalho", "outro"]
                          example: "casa"
      responses:
        "201":
          description: Endereço criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: "450e8400-e29b-41d4-a716-446655440000"
                      id_usuario:
                        type: string
                        format: uuid
                        example: "550e8400-e29b-41d4-a716-446655440000"
                      attributes:
                        type: object
                        properties:
                          logradouro:
                            type: string
                            example: "Rua Exemplo"
                          bairro:
                            type: string
                            example: "Bairro Exemplo"
                          cidade:
                            type: string
                            example: "Cidade Exemplo"
                          estado:
                            type: string
                            example: "UF"
                          pais:
                            type: string
                            example: "País Exemplo"
                          numero:
                            type: integer
                            example: 123
                          complemento:
                            type: string
                            example: "Apto 123"
                          tipo:
                            type: string
                            example: "casa"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_message:
                    type: string
                    example: "Tipo de endereço inválido! Escolha entre casa ou trabalho"
        "404":
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_message:
                    type: string
                    example: "Usuário não encontrado!"
        "409":
          description: Tipo de endereço já existe
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_message:
                    type: string
                    example: "Esse tipo de endereço já existe!"
  /user/{id}/enderecos:
    get:
      tags:
        - Endereços
      summary: Listar endereços de um usuário
      description: Retorna todos os endereços de um usuário específico
      parameters:
        - name: id
          in: path
          required: true
          description: ID do usuário
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lista de endereços
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          example: "450e8400-e29b-41d4-a716-446655440000"
                        id_usuario:
                          type: string
                          format: uuid
                          example: "550e8400-e29b-41d4-a716-446655440000"
                        logradouro:
                          type: string
                          example: "Rua Exemplo"
                        bairro:
                          type: string
                          example: "Bairro Exemplo"
                        cidade:
                          type: string
                          example: "Cidade Exemplo"
                        estado:
                          type: string
                          example: "UF"
                        pais:
                          type: string
                          example: "País Exemplo"
                        numero:
                          type: integer
                          example: 123
                        complemento:
                          type: string
                          example: "Apto 123"
                        tipo:
                          type: string
                          example: "casa"
        "404":
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_message:
                    type: string
                    example: "Usuário não encontrado!"
  /endereco/{id}:
    put:
      tags:
        - Endereços
      summary: Alterar um endereço
      description: Atualiza os dados de um endereço específico
      parameters:
        - name: id
          in: path
          required: true
          description: ID do endereço
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  required:
                    - id_usuario
                    - attributes
                  properties:
                    id_usuario:
                      type: string
                      format: uuid
                      example: "550e8400-e29b-41d4-a716-446655440000"
                    attributes:
                      type: object
                      properties:
                        logradouro:
                          type: string
                          example: "Nova Rua"
                        bairro:
                          type: string
                          example: "Novo Bairro"
                        cidade:
                          type: string
                          example: "Nova Cidade"
                        estado:
                          type: string
                          example: "NE"
                        pais:
                          type: string
                          example: "Novo País"
                        numero:
                          type: integer
                          example: 456
                        complemento:
                          type: string
                          example: "Bloco B"
                        tipo:
                          type: string
                          enum: ["casa", "trabalho", "outro"]
                          example: "trabalho"
      responses:
        "200":
          description: Endereço atualizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: "450e8400-e29b-41d4-a716-446655440000"
                      id_usuario:
                        type: string
                        format: uuid
                        example: "550e8400-e29b-41d4-a716-446655440000"
                      attributes:
                        type: object
                        properties:
                          logradouro:
                            type: string
                            example: "Nova Rua"
                          bairro:
                            type: string
                            example: "Novo Bairro"
                          cidade:
                            type: string
                            example: "Nova Cidade"
                          estado:
                            type: string
                            example: "NE"
                          pais:
                            type: string
                            example: "Novo País"
                          numero:
                            type: integer
                            example: 456
                          complemento:
                            type: string
                            example: "Bloco B"
                          tipo:
                            type: string
                            example: "trabalho"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_message:
                    type: string
                    example: "Tipo de endereço inválido! Escolha entre casa ou trabalho"
        "404":
          description: Endereço não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_message:
                    type: string
                    example: "Endereço não encontrado!"
        "409":
          description: Tipo de endereço já existe
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_message:
                    type: string
                    example: "Esse tipo de endereço já existe!"

    delete:
      tags:
        - Endereços
      summary: Deletar um endereço
      description: Remove um endereço do sistema
      parameters:
        - name: id
          in: path
          required: true
          description: ID do endereço
          schema:
            type: string
            format: uuid
        - name: idUser
          in: query
          required: true
          description: ID do usuário
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Endereço deletado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Endereço deletado com sucesso"
        "404":
          description: Endereço não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_message:
                    type: string
                    example: "Endereço não encontrado!"
...